requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
aiohttp==3.9.1
fake-useragent==1.4.0
python-dotenv==1.0.0
pandas==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3
retry==0.9.2
tqdm==4.66.1
colorlog==6.7.0
python-dateutil==2.8.2
urllib3==2.1.0
certifi==2024.2.2
openpyxl==3.1.2
requests-toolbelt==1.0.0
pyopenssl==23.2.0
